fix : useItems hook으로 typeItems 거르는 코드 뽑음
